ref(metrics): move metrics api metric details endpoint to separate file
